# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RAG-Anything API
  description: 'Multimodal RAG service: query, process documents, insert content.
    Multi-tenant via workspace parameter.'
  version: 1.0.0
servers:
- url: /
  description: Default (relative to base URL)
paths:
  /health:
    get:
      tags:
      - health
      summary: Health
      description: 'Liveness: always 200. Use /ready for readiness.'
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
  /ready:
    get:
      tags:
      - health
      summary: Ready
      description: 'Readiness: 200 if config and API key are present (RAG is created
        per workspace on first use).'
      operationId: ready_ready_get
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
  /config:
    get:
      tags:
      - config
      summary: Config Endpoint
      description: Return non-secret app config (storage types, default workspace,
        parser).
      operationId: config_endpoint_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /config/{workspace_id}:
    get:
      tags:
      - config
      summary: Get Workspace Config
      description: Get config stored for the given workspace (from DB). 404 if not
        set.
      operationId: get_workspace_config_config__workspace_id__get
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - config
      summary: Set Workspace Config
      description: 'Set config for workspace. Payload is partial; missing keys are
        filled from env defaults.

        Only known keys are allowed (validation error for unknown keys). Stored in
        DB and _rag_cache updated.'
      operationId: set_workspace_config_config__workspace_id__post
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceConfigPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /query:
    post:
      tags:
      - query
      summary: Query
      description: Text query over the indexed knowledge base for the given workspace.
      operationId: query_query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /query/multimodal:
    post:
      tags:
      - query
      summary: Query Multimodal
      description: Query with inline multimodal content (tables, equations, image
        paths).
      operationId: query_multimodal_query_multimodal_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryMultimodalRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /content/insert:
    post:
      tags:
      - ingest
      summary: Insert Content
      description: Insert a pre-parsed content list (no file parsing) into the given
        workspace.
      operationId: insert_content_content_insert_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertContentRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /documents/process:
    post:
      tags:
      - ingest
      summary: Process Document
      description: Upload a document and run full RAG processing (parse + index) for
        the given workspace.
      operationId: process_document_documents_process_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_process_document_documents_process_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_process_document_documents_process_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        workspace:
          type: string
          title: Workspace
          description: Tenant/workspace id
          default: default
        output_dir:
          anyOf:
          - type: string
          - type: 'null'
          title: Output Dir
        parse_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Parse Method
        parser:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser
        doc_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Doc Id
      type: object
      required:
      - file
      title: Body_process_document_documents_process_post
    ContentListItem:
      properties:
        type:
          type: string
          title: Type
          description: text, image, table, equation, or custom
        text:
          anyOf:
          - type: string
          - type: 'null'
          title: Text
        img_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Img Path
        image_caption:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Image Caption
        image_footnote:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Image Footnote
        table_body:
          anyOf:
          - type: string
          - type: 'null'
          title: Table Body
        table_caption:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Table Caption
        table_footnote:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Table Footnote
        latex:
          anyOf:
          - type: string
          - type: 'null'
          title: Latex
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        page_idx:
          type: integer
          title: Page Idx
          default: 0
      type: object
      required:
      - type
      title: ContentListItem
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InsertContentRequest:
      properties:
        workspace:
          type: string
          title: Workspace
          description: Tenant/workspace id
          default: default
        content_list:
          items:
            $ref: '#/components/schemas/ContentListItem'
          type: array
          title: Content List
          description: Pre-parsed content items
        file_path:
          type: string
          title: File Path
          description: Reference name for citations
          default: unknown_document
        doc_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Doc Id
        split_by_character:
          anyOf:
          - type: string
          - type: 'null'
          title: Split By Character
        split_by_character_only:
          type: boolean
          title: Split By Character Only
          default: false
      type: object
      required:
      - content_list
      title: InsertContentRequest
    MultimodalItem:
      properties:
        type:
          type: string
          title: Type
          description: 'One of: table, equation, image'
        table_data:
          anyOf:
          - type: string
          - type: 'null'
          title: Table Data
        table_body:
          anyOf:
          - type: string
          - type: 'null'
          title: Table Body
        table_caption:
          anyOf:
          - type: string
          - items:
              type: string
            type: array
          - type: 'null'
          title: Table Caption
        table_footnote:
          anyOf:
          - type: string
          - items:
              type: string
            type: array
          - type: 'null'
          title: Table Footnote
        latex:
          anyOf:
          - type: string
          - type: 'null'
          title: Latex
        equation_caption:
          anyOf:
          - type: string
          - type: 'null'
          title: Equation Caption
        img_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Img Path
        image_caption:
          anyOf:
          - type: string
          - items:
              type: string
            type: array
          - type: 'null'
          title: Image Caption
        image_footnote:
          anyOf:
          - type: string
          - items:
              type: string
            type: array
          - type: 'null'
          title: Image Footnote
      type: object
      required:
      - type
      title: MultimodalItem
    QueryMultimodalRequest:
      properties:
        workspace:
          type: string
          title: Workspace
          description: Tenant/workspace id
          default: default
        query:
          type: string
          title: Query
          description: Question in context of the given multimodal content
        multimodal_content:
          items:
            $ref: '#/components/schemas/MultimodalItem'
          type: array
          title: Multimodal Content
          description: Tables, equations, or image refs
        mode:
          type: string
          title: Mode
          description: Query mode
          default: hybrid
      type: object
      required:
      - query
      - multimodal_content
      title: QueryMultimodalRequest
    QueryRequest:
      properties:
        workspace:
          type: string
          title: Workspace
          description: Tenant/workspace id for data isolation
          default: default
        query:
          type: string
          title: Query
          description: Natural language question
        mode:
          type: string
          title: Mode
          description: 'Query mode: local, global, hybrid, naive, mix, bypass'
          default: hybrid
        vlm_enhanced:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Vlm Enhanced
          description: Use VLM for images in context
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          title: System Prompt
      type: object
      required:
      - query
      title: QueryRequest
    QueryResponse:
      properties:
        answer:
          type: string
          title: Answer
      type: object
      required:
      - answer
      title: QueryResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkspaceConfigPayload:
      properties:
        working_dir:
          anyOf:
          - type: string
          - type: 'null'
          title: Working Dir
        parser:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser
        parse_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Parse Method
        enable_image_processing:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Image Processing
        enable_table_processing:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Table Processing
        enable_equation_processing:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Equation Processing
        kv_storage:
          anyOf:
          - type: string
          - type: 'null'
          title: Kv Storage
        vector_storage:
          anyOf:
          - type: string
          - type: 'null'
          title: Vector Storage
        graph_storage:
          anyOf:
          - type: string
          - type: 'null'
          title: Graph Storage
        doc_status_storage:
          anyOf:
          - type: string
          - type: 'null'
          title: Doc Status Storage
        top_k:
          anyOf:
          - type: integer
          - type: 'null'
          title: Top K
        chunk_top_k:
          anyOf:
          - type: integer
          - type: 'null'
          title: Chunk Top K
        max_entity_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Entity Tokens
        max_relation_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Relation Tokens
        max_total_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Total Tokens
        cosine_better_than_threshold:
          anyOf:
          - type: number
          - type: 'null'
          title: Cosine Better Than Threshold
        vector_db_storage_cls_kwargs:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Vector Db Storage Cls Kwargs
      additionalProperties: false
      type: object
      title: WorkspaceConfigPayload
      description: Allowed workspace config keys for POST /config/{workspace_id}.
        Extra keys are forbidden.
tags:
- name: health
  description: Liveness and readiness probes
- name: config
  description: App and per-workspace config (GET/POST)
- name: query
  description: Text and multimodal RAG queries
- name: ingest
  description: Content insert and document processing